# openapi.yaml
openapi: 3.0.3
info:
  title: Loan Intake Service API
  version: 1.0.0
  description: API for authentication and submitting, reviewing, and deciding loan applications.
servers:
  - url: /
    description: Current server
tags:
  - name: System
  - name: Authentication
  - name: Applications

paths:
  /health:
    get:
      tags: [System]
      summary: Check service health
      operationId: getHealth
      responses:
        "200":
          description: The service is healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"

  /auth/register:
    post:
      tags: [Authentication]
      summary: Register an applicant
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: Applicant registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/MalformedJson"
        "409":
          description: The email is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  /auth/login:
    post:
      tags: [Authentication]
      summary: Log in
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Authentication succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/MalformedJson"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: A new token pair was issued and the presented refresh token was revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/MalformedJson"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke a refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "204":
          description: The token was revoked, or was already unavailable.
        "400":
          $ref: "#/components/responses/MalformedJson"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  /applications:
    post:
      tags: [Applications]
      summary: Submit a loan application
      operationId: submitApplication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplication"
      responses:
        "202":
          description: Application accepted for asynchronous scoring.
          headers:
            Location:
              description: Relative URL of the new application.
              schema:
                type: string
                example: /applications/cm1234567890abcdefghijk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/MalformedJson"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Applications]
      summary: List visible applications
      description: Applicants see their own applications; officers and administrators see all applications.
      operationId: listApplications
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/ApplicationStatus"
      responses:
        "200":
          description: A page of applications.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  /applications/{id}:
    get:
      tags: [Applications]
      summary: Get an application
      operationId: getApplication
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: The requested application.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  /applications/{id}/status:
    patch:
      tags: [Applications]
      summary: Change an application status
      description: Requires an OFFICER or ADMIN access token.
      operationId: updateApplicationStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        $ref: "#/components/requestBodies/StatusChange"
      responses:
        "200":
          description: The updated application.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/MalformedJson"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  /applications/{id}/decision:
    patch:
      tags: [Applications]
      summary: Record an application decision
      description: Requires an OFFICER or ADMIN access token. Uses the same transition rules as the status endpoint.
      operationId: decideApplication
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        $ref: "#/components/requestBodies/StatusChange"
      responses:
        "200":
          description: The updated application.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/MalformedJson"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/InvalidTransition"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ApplicationId:
      name: id
      in: path
      required: true
      description: CUID2 application identifier.
      schema:
        type: string
        minLength: 1
      example: cm1234567890abcdefghijk
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ApplicationStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ApplicationStatus"

  requestBodies:
    StatusChange:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StatusChange"

  responses:
    MalformedJson:
      description: The request body is not valid JSON.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated user cannot perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The application does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidTransition:
      description: The requested status transition is not allowed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: One or more request values failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: More than five attempts were made from this IP within fifteen minutes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    Health:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    Role:
      type: string
      enum: [APPLICANT, OFFICER, ADMIN]

    User:
      type: object
      additionalProperties: false
      required: [id, email, role]
      properties:
        id:
          type: string
          example: cm1234567890abcdefghijk
        email:
          type: string
          format: email
          example: applicant@example.com
        role:
          $ref: "#/components/schemas/Role"

    Credentials:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: applicant@example.com
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          example: correct-horse-battery-staple

    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 64
          maxLength: 64
          description: Opaque 64-character refresh token.

    TokenPair:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, tokenType, expiresAt]
      properties:
        accessToken:
          type: string
          description: JWT access token.
        refreshToken:
          type: string
          minLength: 64
          maxLength: 64
        tokenType:
          type: string
          enum: [Bearer]
        expiresAt:
          type: integer
          description: Access-token lifetime in seconds.
          example: 900

    LoginResponse:
      allOf:
        - $ref: "#/components/schemas/TokenPair"
        - type: object
          required: [user]
          properties:
            user:
              $ref: "#/components/schemas/User"

    CreateApplication:
      type: object
      additionalProperties: false
      required: [amount, term, monthlyIncome, purpose]
      properties:
        amount:
          type: number
          exclusiveMinimum: true
          minimum: 0
          maximum: 100000000
          example: 250000
        term:
          type: integer
          minimum: 6
          maximum: 360
          description: Loan term in months.
          example: 24
        monthlyIncome:
          type: number
          exclusiveMinimum: true
          minimum: 0
          maximum: 100000000
          example: 45000
        purpose:
          type: string
          minLength: 3
          maxLength: 140
          example: Home improvements

    ApplicationStatus:
      type: string
      enum: [PENDING, IN_REVIEW, APPROVED, REJECTED, NEEDS_REVIEW]

    Application:
      type: object
      additionalProperties: false
      required:
        - id
        - applicantId
        - amount
        - term
        - monthlyIncome
        - purpose
        - status
        - score
        - decidedAt
        - createdAt
      properties:
        id:
          type: string
          example: cm1234567890abcdefghijk
        applicantId:
          type: string
          example: cm0987654321abcdefghijk
        amount:
          type: string
          pattern: "^[0-9]+\\.[0-9]{2}$"
          description: Decimal amount serialized with two fraction digits.
          example: "250000.00"
        term:
          type: integer
          example: 24
        monthlyIncome:
          type: string
          pattern: "^[0-9]+\\.[0-9]{2}$"
          description: Decimal income serialized with two fraction digits.
          example: "45000.00"
        purpose:
          type: string
          example: Home improvements
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        score:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        decidedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    ApplicationPage:
      type: object
      additionalProperties: false
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0

    StatusChange:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status:
          type: string
          enum: [IN_REVIEW, APPROVED, REJECTED]
        reason:
          type: string
          minLength: 3
          maxLength: 500
          example: Applicant documents verified

    ErrorDetail:
      type: object
      additionalProperties: true
      description: Optional structured context, such as validation issues or transition states.

    Error:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum: [NOT_FOUND, UNAUTHORIZED, FORBIDDEN, CONFLICT, VALIDATION_ERROR, RATE_LIMIT_EXCEEDED, INTERNAL_ERROR]
        message:
          type: string
        details:
          $ref: "#/components/schemas/ErrorDetail"
        requestId:
          type: string

    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"